+++ backend-node/CHANGELOG.md
@@ -13,3 +13,4 @@
 * Cascade rounding
 * Open route for compute bills
 * Update cascade rounding algorithm to compute complex cases
+* Add prometheus client for metrics support